requirements.txt
alembic==1.8.1
greenlet==2.0.2
Mako==1.2.4
MarkupSafe==2.1.3
mqtt==0.0.1
mysql-connector-python==8.1.0
mysqlclient==2.1.1
paho-mqtt==1.6.1
protobuf==4.21.12
python-dotenv==0.21.0
SQLAlchemy==1.4.44
